src/apps/auth/pyproject.toml
[tool.poetry]
name = "auth"
version = "0.1.0"
description = "Microservice of auth"
authors = ["Akuev Anton"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
alembic = "^1.13.2"
asyncpg = "^0.29.0"
pyjwt = {extras = ["crypto"], version = "^2.9.0"}
pydantic-settings = "^2.4.0"
psycopg2-binary = "^2.9.9"
bcrypt = "^4.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
